Create sql_census_analysis
This is a section of a series of SQL queries I created to find the population density in each U.S. census tract. The purpose of this project was to create a demographic profile for each census tract in Minnesota and compare the demographics and voting patterns between high density areas and low density areas. The data pulled from this query was specifically used to create a density map layer on ArcGIS.